Sửa lỗi chức năng hiển thị sản phẩm và danh mục

+++ src/main/java/com/cdweb/Treestore/domain/Tree.java
@@ -42,7 +42,7 @@ public class Tree {
     private String quantity;
 
     @Column(name = "discount")
-    private Float discount;
+    private double discount;
 
     @Column(name = "hot_tree")
     private Boolean hotTree;
